backend/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sync-backend"
version = "0.1.0"
description = "Backend service for the Sync audiobook-EPUB alignment project."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "celery>=5.4,<6.0",
  "fastapi>=0.115,<1.0",
  "orjson>=3.10,<4.0",
  "pydantic-settings>=2.7,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "sqlalchemy>=2.0,<3.0",
  "structlog>=24.4,<25.0",
  "uvicorn[standard]>=0.32,<1.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "mypy>=1.13,<2.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.8,<1.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
pretty = true
mypy_path = "src"
packages = ["sync_backend"]

[[tool.mypy.overrides]]
module = ["celery", "celery.*"]
ignore_missing_imports = true
